+++ lib/src/features/analysis/data/data_sources/sets/state_class_instance_set.dart
@@ -1,9 +1,0 @@
-import 'package:analyzer/dart/ast/ast.dart';
-
-typedef StateClassInstance = ({
-  String instanceId,
-  String filePath,
-  String stateClassName,
-  ClassDeclaration classDeclaration,
-  MethodDeclaration? buildMethod,
-});